Require accountId config setting to avoid mess
The API permission token can be setup to only work on
a single account. However, at times that has either been
configured wrong (or Cloudflare did not properly enforce
the configuration), and all domains across all owned accounts
showed up...

Consequently, pinning this to a single account means we avoid
lots of "where should this go?" questions.

Multiple account configuration could be viable in the future, but
really is not worth the overhead at this point...